dedupe wheel-extras regex into libs/patchwheel
The pip-extras split regex was duplicated in bundle/libraries and in
libs/testserver/jobs.go. libs/ must not import bundle/, so move the
shared helper down into libs/patchwheel (already the wheel-parsing home
and already a dependency of bundle/libraries) as SplitWheelExtras, and
have both call sites use it. testserver only needs the base path, so it
discards the extras return.
