+++ bundle/libraries/expand_glob_references.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"path/filepath"
-	"regexp"
 	"strings"
 
 	"github.com/databricks/cli/bundle"
@@ -14,20 +13,6 @@ import (
 )
 
 type expand struct{}
-
-// Matches a pip extras suffix right after ".whl", e.g. "[train]" in "pkg.whl[train]".
-// Anchored so glob classes elsewhere (e.g. "pkg[12].whl") are left alone.
-var wheelExtrasRegex = regexp.MustCompile(`(?i)\.whl(\[[^\]]*\])$`)
-
-// splitWheelExtras splits a wheel path into its base and pip extras suffix (e.g. "[train]").
-func splitWheelExtras(path string) (string, string) {
-	m := wheelExtrasRegex.FindStringSubmatch(path)
-	if m == nil {
-		return path, ""
-	}
-	extras := m[1]
-	return strings.TrimSuffix(path, extras), extras
-}
 
 func matchError(p dyn.Path, l []dyn.Location, message string) diag.Diagnostic {
 	return diag.Diagnostic{
@@ -132,7 +117,7 @@ func expandEnvironmentDeps(ctx context.Context, b *bundle.Bundle, p dyn.Path, v 
 		}
 
 		// Strip extras before globbing so "[...]" isn't read as a glob class, then re-append.
-		path, extras := splitWheelExtras(path)
+		path, extras := patchwheel.SplitWheelExtras(path)
 
 		matches, err := findMatches(ctx, b, path)
 		if err != nil {
